refactor: CR clean ups
