--- src/OsAuth/index.ts
@@ -1,2 +1,0 @@
-export * from './OsAuthScreen'
-export * from './OsLoginScreen'
