+++ src/Settings/ApplicationSettings/ApplicationSettingsScreen.tsx
@@ -3,9 +3,9 @@ import {defineMessages, useIntl} from 'react-intl'
 import {ScrollView, StyleSheet, Switch} from 'react-native'
 import DeviceInfo from 'react-native-device-info'
 
-import {useAuthOsEnabled, useAuthOsError, useAuthWithOs} from '../../auth'
+import {useAuthOsEnabled, useAuthOsError, useAuthSetting, useAuthWithOs} from '../../auth'
 import {StatusBar} from '../../components'
-import {useAuthSetting, useCrashReports} from '../../hooks'
+import {useCrashReports} from '../../hooks'
 import globalMessages from '../../i18n/global-messages'
 import {CONFIG, isNightly} from '../../legacy/config'
 import {useWalletNavigation} from '../../navigation'
