+++ src/Settings/WalletSettings/WalletSettingsScreen.tsx
@@ -4,10 +4,10 @@ import {defineMessages, useIntl} from 'react-intl'
 import {InteractionManager, ScrollView, StyleSheet, Switch} from 'react-native'
 import {UseMutationOptions} from 'react-query'
 
-import {useEasyConfirmationEnabled} from '../../auth'
+import {useAuthSetting, useEasyConfirmationEnabled} from '../../auth'
 import {useAuth} from '../../auth/AuthProvider'
 import {StatusBar} from '../../components'
-import {useAuthSetting, useCloseWallet, useResync, useWalletName} from '../../hooks'
+import {useCloseWallet, useResync, useWalletName} from '../../hooks'
 import {confirmationMessages} from '../../i18n/global-messages'
 import {DIALOG_BUTTONS, showConfirmationDialog} from '../../legacy/actions'
 import {isByron, isHaskellShelley} from '../../legacy/config'
