+++ src/Settings/index.ts
@@ -1,2 +1,1 @@
 export * from './SettingsScreenNavigator'
-export * from './types'
