+++ src/Storage/index.ts
@@ -1,1 +1,2 @@
+export * from './migrations/hooks'
 export * from './StorageProvider'
