+++ src/auth/CheckPinInput/CheckPinInput.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import {defineMessages, useIntl} from 'react-intl'
 
-import {useCheckPin} from '../../hooks'
 import {errorMessages} from '../../i18n/global-messages'
 import {showErrorDialog} from '../../legacy/actions'
 import {CONFIG} from '../../legacy/config'
 import {useStorage} from '../../Storage'
+import {useCheckPin} from '../hooks'
 import {PinInput, PinInputRef} from '../PinInput'
 
 export const CheckPinInput = ({onValid}: {onValid: () => void}) => {
