+++ src/auth/PinLoginScreen/PinLoginScreen.tsx
@@ -3,12 +3,12 @@ import {defineMessages, useIntl} from 'react-intl'
 import {SafeAreaView} from 'react-native-safe-area-context'
 
 import {StatusBar} from '../../components'
-import {useCheckPin} from '../../hooks'
 import {errorMessages} from '../../i18n/global-messages'
 import {showErrorDialog} from '../../legacy/actions'
 import {CONFIG} from '../../legacy/config'
 import {useStorage} from '../../Storage'
 import {useAuth} from '../AuthProvider'
+import {useCheckPin} from '../hooks'
 import {PinInput, PinInputRef} from '../PinInput'
 
 export const PinLoginScreen = () => {
