+++ src/auth/index.ts
@@ -1,9 +1,11 @@
-export * from './authOS'
 export * from './backgroundTimeout'
 export * from './ChangePinScreen'
 export * from './CheckPinInput'
 export * from './CreatePinInput'
 export * from './CreatePinScreen'
 export * from './EncryptedStorage'
+export * from './hooks'
+export * from './OsAuthScreen'
+export * from './OsLoginScreen'
 export * from './PinInput'
 export * from './PinLoginScreen'
