+++ src/i18n/LanguageProvider.tsx
@@ -4,6 +4,7 @@ import {IntlProvider} from 'react-intl'
 import {NativeModules, Platform, Text} from 'react-native'
 import {useMutation, UseMutationOptions, useQuery, useQueryClient, UseQueryOptions} from 'react-query'
 
+import crashReporting from '../legacy/crashReporting'
 import {isEmptyString} from '../legacy/utils'
 import {updateLanguageSettings} from '.'
 import {supportedLanguages} from './languages'
@@ -44,6 +45,7 @@ const useLanguageCode = ({onSuccess, ...options}: UseQueryOptions<string> = {}) 
       return defaultLanguageCode
     },
     onSuccess: (languageCode) => {
+      crashReporting.setStringValue('language_code', languageCode)
       updateLanguageSettings(languageCode)
       onSuccess?.(languageCode)
     },
